:recycle: Refactor JP2WSIReader
