+++ tests/test_slide_thumbnail.py
@@ -10,6 +10,7 @@ from click.testing import CliRunner
 from tiatoolbox import cli
 from tiatoolbox.utils.env_detection import running_on_ci
 from tiatoolbox.wsicore import wsireader
+from tiatoolbox.wsicore.wsireader import jp2
 
 
 def test_wsireader_get_thumbnail_openslide(sample_svs: Path) -> None:
@@ -22,7 +23,7 @@ def test_wsireader_get_thumbnail_openslide(sample_svs: Path) -> None:
 
 def test_wsireader_get_thumbnail_jp2(sample_jp2: Path) -> None:
     """Test for get_thumbnail as a python function."""
-    wsi = wsireader.JP2WSIReader(sample_jp2)
+    wsi = jp2.JP2WSIReader(sample_jp2)
     slide_thumbnail = wsi.slide_thumbnail()
     assert isinstance(slide_thumbnail, np.ndarray)
     assert slide_thumbnail.dtype == "uint8"
