+++ tiatoolbox/wsicore/wsireader/__init__.py
@@ -9,7 +9,6 @@ from .base import (
     AnnotationStoreReader,
     ArrayView,
     FsspecJsonWSIReader,
-    JP2WSIReader,
     NGFFWSIReader,
     OpenSlideWSIReader,
     TIFFWSIReader,
@@ -24,6 +23,7 @@ from .detection import is_dicom, is_ngff, is_tiled_tiff, is_url, is_zarr
 
 # Import reader classes
 from .dicom import DICOMWSIReader
+from .jp2 import JP2WSIReader
 
 __all__ = [
     "AnnotationStoreReader",
